airflow/refactored/requirements.txt
python-dotenv
psycopg2-binary
requests
beautifulsoup4
chardet
boto3